DOSIS-1045 Move application code to functions directory
