--- services/gp-search/application/services/ftrs_service/fhir_mapper/__init__.py
@@ -1,3 +1,0 @@
-from application.services.ftrs_service.fhir_mapper.bundle_mapper import BundleMapper
-
-__all__ = ["BundleMapper"]
