+++ services/gp-search/functions/ftrs_service/fhir_mapper/__init__.py
@@ -1,0 +1,3 @@
+from functions.ftrs_service.fhir_mapper.bundle_mapper import BundleMapper
+
+__all__ = ["BundleMapper"]
