+++ services/gp-search/functions/ftrs_service/fhir_mapper/organization_mapper.py
@@ -2,7 +2,8 @@ from fhir.resources.R4B.address import Address
 from fhir.resources.R4B.contactpoint import ContactPoint
 from fhir.resources.R4B.identifier import Identifier
 from fhir.resources.R4B.organization import Organization
-from services.ftrs_service.repository.dynamo import (
+
+from functions.ftrs_service.repository.dynamo import (
     OrganizationRecord,
     OrganizationValue,
 )
