+++ services/gp-search/functions/ftrs_service/repository/dynamo.py
