+++ services/gp-search/pyproject.toml
@@ -52,7 +52,7 @@ line-ending = "auto"
 docstring-code-format = true
 
 [tool.pytest.ini_options]
-pythonpath = "application"
+pythonpath = "functions"
 python_files = "test_*.py"
 testpaths = "tests"
 addopts = "-v"
